Move ErrorText to flat structure
