+++ packages/client/src/components/form/DocumentUploadfield/SimpleDocumentUploader.tsx
@@ -10,7 +10,7 @@
  * graphic logo are (registered/a) trademark(s) of Plan International.
  */
 import { ImageUploader } from '@opencrvs/components/lib/ImageUploader'
-import { ErrorText } from '@opencrvs/components/lib/forms/ErrorText'
+import { ErrorText } from '@opencrvs/components/lib/ErrorText'
 import { DocumentPreview } from '@client/components/form/DocumentUploadfield/DocumentPreview'
 import { IFormFieldValue, IAttachmentValue } from '@client/forms'
 import * as React from 'react'
