+++ packages/client/src/views/SelectVitalEvent/SelectVitalEvent.tsx
@@ -14,7 +14,7 @@ import styled, { keyframes } from '@client/styledComponents'
 import { WrappedComponentProps as IntlShapeProps, injectIntl } from 'react-intl'
 import { connect } from 'react-redux'
 import { PrimaryButton } from '@opencrvs/components/lib/buttons'
-import { ErrorText } from '@opencrvs/components/lib/forms/ErrorText'
+import { ErrorText } from '@opencrvs/components/lib/ErrorText'
 import {
   FixedEventTopBar,
   RadioButton
